+++ pkg/query/match.go
@@ -4,6 +4,6 @@ type Match struct {
 	Groupname string `json:"group"`
 	Series    string `json:"series"`
 	Smooth    int    `json:"smooth"`
-	StartSeq  int64  `json:"startSeq"`
-	EndSeq    int64  `json:"endSeq"`
+	StartSeq  int32  `json:"startSeq"`
+	EndSeq    int32  `json:"endSeq"`
 }
